Backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
torch
pydantic
moviepy
openai-whisper
transformers[torch]
BitsAndBytes
fastapi
python-multipart
uvicorn
asyncio
aiofiles
ffmpeg-python